http://web.archive.org/web/20070505061921/http://www.djangosnippets.org/snippets/214/
WTForm is an extension to the django newforms library allowing the developer, in a very flexible way, to layout the form fields using fieldsets and columns

WTForm was built with the well-documented YUI Grid CSS in mind when rendering the columns and fields. This should make it easy to implement WTForm in your own applications.